QFIT without vana_calibration
Modified the 'qfit' and the 'qfit_plot' methods of 'Data' to works without vana_calibration info, i.e. when the object is created, no 'vana_filename' is passed and therefore no information on a previously fitted vanadium sample are available for the calibration.
This allow to perform the fit of the Vanadium.